remove dispatcher
