--- workers/unity/Packages/com.improbable.gdk.core/Dispatcher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd41a9a21fb0bdd4c8de5659511e7e05
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
